refactor(internal): make smart_search a package
Move the flat _search_*.py siblings into a smart_search/ package with
private submodules, so the leading-underscore naming reads correctly as
package internals and matches the repo's plain-named service modules.

- smart_search.py -> smart_search/__init__.py (public shell)
- _search_{base,config,deep,overview,entities}.py -> smart_search/_{...}.py
- Relative imports re-depthed; intra-package imports use single dot
- Public path ha_mcp.tools.smart_search unchanged; get_global_settings
  patch target preserved; AUTOMATION_CONFIG_TIME_BUDGET test + docstring
  harvester paths repointed to smart_search/_deep|_overview|_entities
- AGENTS.md module-size note updated to the package example
