+++ src/ha_mcp/tools/smart_search/_config.py
