+++ src/ha_mcp/tools/smart_search/_deep.py
@@ -10,9 +10,13 @@ from typing import Any
 from fastmcp import Context
 from fastmcp.exceptions import ToolError
 
-from ..utils.fuzzy_search import BM25Scorer, calculate_ratio, tokenize
-from ._search_base import _SearchBase
-from ._search_config import (
+from ...utils.fuzzy_search import BM25Scorer, calculate_ratio, tokenize
+from ..helpers import exception_to_structured_error, safe_info, safe_progress
+from ..tools_config_dashboards import fetch_dashboards_list
+from ..tools_config_entry_flow import FLOW_HELPER_TYPES
+from ..tools_integrations import fetch_entry_options
+from ._base import _SearchBase
+from ._config import (
     AUTOMATION_CONFIG_TIME_BUDGET,
     BULK_REST_TIMEOUT,
     BULK_WEBSOCKET_TIMEOUT,
@@ -22,10 +26,6 @@ from ._search_config import (
     SCENE_CONFIG_TIME_BUDGET,
     SCRIPT_CONFIG_TIME_BUDGET,
 )
-from .helpers import exception_to_structured_error, safe_info, safe_progress
-from .tools_config_dashboards import fetch_dashboards_list
-from .tools_config_entry_flow import FLOW_HELPER_TYPES
-from .tools_integrations import fetch_entry_options
 
 logger = logging.getLogger(__name__)
